+++ src/Validator/AbstractBetweenParser.php
@@ -1,23 +1,25 @@
 <?php
 
-namespace MPScholten\RequestParser\ValidationParser;
+namespace MPScholten\RequestParser\Validator;
 
 use MPScholten\RequestParser\AbstractValueParser;
 use MPScholten\RequestParser\Config;
 
-abstract class AbstractSmallerThanOrEqualToParser extends AbstractValueParser
+abstract class AbstractBetweenParser extends AbstractValueParser
 {
+    protected $minValue;
     protected $maxValue;
 
-    public function __construct(Config $config, $name, $value, $maxValue)
+    public function __construct(Config $config, $name, $value, $minValue, $maxValue)
     {
+        $this->minValue = $minValue;
         $this->maxValue = $maxValue;
         parent::__construct($config, $name, $value);
     }
 
     protected function parse($value)
     {
-        if ($value <= $this->maxValue) {
+        if ($value >= $this->minValue && $value <= $this->maxValue) {
             return $value;
         }
         return null;
