+++ src/Validator/EmailParser.php
@@ -1,19 +1,19 @@
 <?php
 
-namespace MPScholten\RequestParser\ValidationParser;
+namespace MPScholten\RequestParser\Validator;
 
 use MPScholten\RequestParser\AbstractValueParser;
 
-class UrlParser extends AbstractValueParser
+class EmailParser extends AbstractValueParser
 {
     protected function describe()
     {
-        return "a valid URL";
+        return "a valid email address";
     }
 
     protected function parse($value)
     {
-        if (filter_var($value, FILTER_VALIDATE_URL) === false) {
+        if (filter_var($value, FILTER_VALIDATE_EMAIL) === false) {
             return null;
         }
         return (string) $value;
