+++ tests/TypeSpecTest.php
@@ -11,24 +11,24 @@ use MPScholten\RequestParser\CommaSeparatedStringParser;
 use MPScholten\RequestParser\CommaSeparatedYesNoBooleanParser;
 use MPScholten\RequestParser\Config;
 use MPScholten\RequestParser\DateTimeParser;
-use MPScholten\RequestParser\ValidationParser\EmailParser;
-use MPScholten\RequestParser\ValidationParser\FloatBetweenParser;
-use MPScholten\RequestParser\ValidationParser\IntBetweenParser;
+use MPScholten\RequestParser\Validator\EmailParser;
+use MPScholten\RequestParser\Validator\FloatBetweenParser;
+use MPScholten\RequestParser\Validator\IntBetweenParser;
 use MPScholten\RequestParser\IntParser;
 use MPScholten\RequestParser\FloatParser;
 use MPScholten\RequestParser\TrimParser;
-use MPScholten\RequestParser\ValidationParser\UrlParser;
-use MPScholten\RequestParser\ValidationParser\FloatLargerThanOrEqualToParser;
-use MPScholten\RequestParser\ValidationParser\FloatLargerThanParser;
-use MPScholten\RequestParser\ValidationParser\FloatSmallerThanOrEqualToParser;
-use MPScholten\RequestParser\ValidationParser\FloatSmallerThanParser;
-use MPScholten\RequestParser\ValidationParser\IntLargerThanParser;
-use MPScholten\RequestParser\ValidationParser\IntSmallerThanOrEqualToParser;
-use MPScholten\RequestParser\ValidationParser\IntSmallerThanParser;
+use MPScholten\RequestParser\Validator\UrlParser;
+use MPScholten\RequestParser\Validator\FloatLargerThanOrEqualToParser;
+use MPScholten\RequestParser\Validator\FloatLargerThanParser;
+use MPScholten\RequestParser\Validator\FloatSmallerThanOrEqualToParser;
+use MPScholten\RequestParser\Validator\FloatSmallerThanParser;
+use MPScholten\RequestParser\Validator\IntLargerThanParser;
+use MPScholten\RequestParser\Validator\IntSmallerThanOrEqualToParser;
+use MPScholten\RequestParser\Validator\IntSmallerThanParser;
 use MPScholten\RequestParser\YesNoBooleanParser;
 use MPScholten\RequestParser\BooleanParser;
 use MPScholten\RequestParser\JsonParser;
-use MPScholten\RequestParser\ValidationParser\OneOfParser;
+use MPScholten\RequestParser\Validator\OneOfParser;
 use MPScholten\RequestParser\StringParser;
 use MPScholten\RequestParser\TypeParser;
